backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
pymongo
motor
langchain
langchain-community
langchain-openai
langchain-huggingface
sentence-transformers
pinecone-client
chromadb
faiss-cpu
beautifulsoup4
requests
python-docx
pypdf
python-dotenv
httpx
